+++ src/App.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from "react";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
-import TodoList from "./TodoList";
-import AddTodoForm from "./AddTodoForm";
+import TodoList from "./components/TodoList";
+import AddTodoForm from "./components/AddTodoForm";
 import { FaRegCalendarAlt } from "react-icons/fa";
 import { TbArrowsSort } from "react-icons/tb";
 import "./App.css"
